Move JsonRequired annotation and its TypeAdapterFactory to internal package
