+++ auth0/src/main/java/com/auth0/android/request/internal/GsonProvider.kt
@@ -3,7 +3,6 @@ package com.auth0.android.request.internal
 import androidx.annotation.VisibleForTesting
 import com.auth0.android.result.Credentials
 import com.auth0.android.result.UserProfile
-import com.auth0.android.util.JsonRequiredTypeAdapterFactory
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.google.gson.reflect.TypeToken
