+++ auth0/src/main/java/com/auth0/android/request/internal/JsonRequired.kt
@@ -1,0 +1,5 @@
+package com.auth0.android.request.internal
+
+@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FIELD)
+internal annotation class JsonRequired
