--- auth0/src/main/java/com/auth0/android/util/JsonRequired.java
@@ -1,11 +1,0 @@
-package com.auth0.android.util;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface JsonRequired {
-}
